+++ package/src/hook.ts
